Refactor PSP and Vita guides.
This commit refactors our PSP and Vita guides to be fully separate with a console distinction, there is now one guide for PSP consoles, and one guide for Vita consoles, regardless of the game being patched.

The PSP guide has been updated to show the much easier browser-based patching method. It is my intent to not show the Refresher-based patching method to users due to its relative complexity.

The Vita guide has been updated to showcase the steps for LBP PSP (under Adrenaline) and LBP Vita. This is due to AutoPlugin2 making the process of patching these two games very similar.

`#how-to-connect` has been updated accordingly to push the users towards the right method for their console.